Adding iiQKA.OS2 support (#228)
* Add file structure change in krl folder

* Change kuka_kss_rsi_driver to kuka_rsi_driver

* Adding documentation to iiQKA.OS2

* test bulet ponts

* Change bulet point structure

* Edit features

* fix xml

* Fix code spell and white space

* fix misspelling

* fix whitespace

* change folder name

* change markdown for folder name change

* class description change

* fix for xml encoding

* fix new line error

* Update kuka_rsi_driver/hardware_interface.xml


* Update doc/wiki/3_iiQKA.OS2_RSI.md


* Update doc/wiki/3_iiQKA.OS2_RSI.md


* Update doc/wiki/3_iiQKA.OS2_RSI.md


* fix client ip config doc

---------
