+++ kuka_rsi_driver/hardware_interface.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+<library path="kuka_rsi_driver">
+  <class type="kuka_rsi_driver::KukaRSIHardwareInterface"
+         base_class_type="hardware_interface::SystemInterface">
+    <description>ros2_control hardware interface for RSI-controlled robots</description>
+  </class>
+</library>
