+++ kuka_rsi_driver/krl/KSS/deprecated/ros_rsi.rsi
