+++ kuka_rsi_driver/krl/iiQKA_OS2/ros_rsi_ethernet.xml
@@ -1,0 +1,35 @@
+<ROOT>
+   <CONFIG>
+      <IP_NUMBER>ip.address.of.rospc</IP_NUMBER>   <!-- IP-number of the external socket -->
+      <PORT>59152</PORT>                   <!-- Port-number of the external socket -->
+      <SENTYPE>KROSHU</SENTYPE>        <!-- The name of your system send in <Sen Type="" > -->
+      <ONLYSEND>FALSE</ONLYSEND>       <!-- TRUE means the client don't expect answers. Do not send anything to robot -->
+   </CONFIG>
+   <!-- RSI Data: TYPE=  "BOOL", "STRING", "LONG", "DOUBLE" -->
+   <!-- INDX= "INTERNAL" switch on internal read values. Needed by DEF_... -->
+   <!-- INDX= "nmb" Input/Output index of RSI-Object / Maximum of RSI Channels: 64  -->
+   <!-- HOLDON="1", set this output index of RSI Object to the last value  -->
+   <!-- DEF_Delay count the late packages and send it back to server  -->
+   <!-- DEF_Tech: .C = advance .T = main run / .C1 advance set function generator 1 -->
+
+   <SEND>
+      <ELEMENTS>
+         <ELEMENT TAG="DEF_RIst" TYPE="DOUBLE" INDX="INTERNAL" />
+         <ELEMENT TAG="DEF_RSol" TYPE="DOUBLE" INDX="INTERNAL" />
+         <ELEMENT TAG="DEF_AIPos" TYPE="DOUBLE" INDX="INTERNAL" />
+         <ELEMENT TAG="DEF_ASPos" TYPE="DOUBLE" INDX="INTERNAL" />
+         <ELEMENT TAG="DEF_Delay" TYPE="LONG" INDX="INTERNAL" />
+      </ELEMENTS>
+   </SEND>
+   <RECEIVE>
+      <ELEMENTS>
+         <ELEMENT TAG="AK.A1" TYPE="DOUBLE" INDX="1" HOLDON="1" />
+         <ELEMENT TAG="AK.A2" TYPE="DOUBLE" INDX="2" HOLDON="1" />
+         <ELEMENT TAG="AK.A3" TYPE="DOUBLE" INDX="3" HOLDON="1" />
+         <ELEMENT TAG="AK.A4" TYPE="DOUBLE" INDX="4" HOLDON="1" />
+         <ELEMENT TAG="AK.A5" TYPE="DOUBLE" INDX="5" HOLDON="1" />
+         <ELEMENT TAG="AK.A6" TYPE="DOUBLE" INDX="6" HOLDON="1" />
+		   <ELEMENT TAG="Stop" TYPE="BOOL" INDX="7" HOLDON="0" />
+      </ELEMENTS>
+   </RECEIVE>
+</ROOT>
