+++ kuka_rsi_driver/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kuka_kss_rsi_driver</name>
+  <name>kuka_rsi_driver</name>
   <version>0.9.2</version>
   <description>A ROS2 hardware interface for use with KUKA RSI</description>
   <maintainer email="svastits1@gmail.com">Aron Svastits</maintainer>
