+++ kuka_rsi_driver/src/robot_manager_node.cpp
@@ -16,7 +16,7 @@
 #include "communication_helpers/service_tools.hpp"
 
 #include "kuka_drivers_core/controller_names.hpp"
-#include "kuka_kss_rsi_driver/robot_manager_node.hpp"
+#include "kuka_rsi_driver/robot_manager_node.hpp"
 
 using namespace controller_manager_msgs::srv;  // NOLINT
 using namespace lifecycle_msgs::msg;           // NOLINT
